ml-engine/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
sqlalchemy
numpy
scipy
pytest
httpx